Fix additional-overlays example
