--- examples/additional-overlays/data/cities.json
@@ -1,22 +1,0 @@
-[
-  {"city":"New York","state":"New York","latitude":40.6643,"longitude":-73.9385},
-  {"city":"Los Angeles","state":"California","latitude":34.0194,"longitude":-118.4108},
-  {"city":"Chicago","state":"Illinois","latitude":41.8376,"longitude":-87.6818},
-  {"city":"Houston","state":"Texas","latitude":29.7805,"longitude":-95.3863},
-  {"city":"Phoenix","state":"Arizona","latitude":33.5722,"longitude":-112.0880},
-  {"city":"Philadelphia","state":"Pennsylvania","latitude":40.0094,"longitude":-75.1333},
-  {"city":"San Antonio","state":"Texas","latitude":29.4724,"longitude":-98.5251},
-  {"city":"San Diego","state":"California","latitude":32.8153,"longitude":-117.1350},
-  {"city":"Dallas","state":"Texas","latitude":32.7757,"longitude":-96.7967},
-  {"city":"San Jose","state":"California","latitude":37.2969,"longitude":-121.8193},
-  {"city":"Austin","state":"Texas","latitude":30.3072,"longitude":-97.7560},
-  {"city":"Jacksonville","state":"Florida","latitude":30.3370,"longitude":-81.6613},
-  {"city":"San Francisco","state":"California","latitude":37.7751,"longitude":-122.4193},
-  {"city":"Columbus","state":"Ohio","latitude":39.9848,"longitude":-82.9850},
-  {"city":"Indianapolis","state":"Indiana","latitude":39.7767,"longitude":-86.1459},
-  {"city":"Fort Worth","state":"Texas","latitude":32.7795,"longitude":-97.3463},
-  {"city":"Charlotte","state":"North Carolina","latitude":35.2087,"longitude":-80.8307},
-  {"city":"Seattle","state":"Washington","latitude":47.6205,"longitude":-122.3509},
-  {"city":"Denver","state":"Colorado","latitude":39.7618,"longitude":-104.8806},
-  {"city":"El Paso","state":"Texas","latitude":31.8484,"longitude":-106.4270}
-]
