--- examples/additional-overlays/data/routes-example.json
@@ -1,274 +1,0 @@
-[
-  {
-    "coordinates": [
-      [
-        -122.3933557873733,
-        37.794219992184324
-      ],
-      [
-        -122.39367189820892,
-        37.79398165755795
-      ],
-      [
-        -122.39401838092769,
-        37.79388241198406
-      ],
-      [
-        -122.3947071294908,
-        37.794423228499795
-      ],
-      [
-        -122.39494688373745,
-        37.79445423060049
-      ],
-      [
-        -122.39571409732676,
-        37.793858299053696
-      ],
-      [
-        -122.39741853206209,
-        37.79245284608079
-      ],
-      [
-        -122.39918399515115,
-        37.79106459055073
-      ],
-      [
-        -122.40003607642596,
-        37.79035713167143
-      ],
-      [
-        -122.40213357823212,
-        37.790068858430104
-      ],
-      [
-        -122.40388909604812,
-        37.78983463559349
-      ],
-      [
-        -122.4054166245374,
-        37.78960041201444
-      ],
-      [
-        -122.40869967084274,
-        37.78920403195862
-      ],
-      [
-        -122.41193711928267,
-        37.788789632353996
-      ],
-      [
-        -122.41526576345333,
-        37.788429282982534
-      ],
-      [
-        -122.41848041296066,
-        37.78799686141706
-      ],
-      [
-        -122.42820266442968,
-        37.78667030140305
-      ],
-      [
-        -122.43987587617389,
-        37.785191419954316
-      ],
-      [
-        -122.43880657433475,
-        37.78061374219536
-      ],
-      [
-        -122.43827192341516,
-        37.777866999452826
-      ],
-      [
-        -122.44669267539851,
-        37.776704885260315
-      ],
-      [
-        -122.44660356691189,
-        37.7757540509692
-      ],
-      [
-        -122.45453422221897,
-        37.77476798767644
-      ]
-    ]
-  },
-  {
-    "coordinates": [
-      [
-        -122.44080380201595,
-        37.75251037780083
-      ],
-      [
-        -122.4425396821639,
-        37.7523902843234
-      ],
-      [
-        -122.44225217701438,
-        37.75281918870459
-      ],
-      [
-        -122.44200806886859,
-        37.75321806754765
-      ],
-      [
-        -122.44180193310093,
-        37.753471118602434
-      ],
-      [
-        -122.44130829218389,
-        37.75388286084137
-      ],
-      [
-        -122.44084177439416,
-        37.75459482638962
-      ],
-      [
-        -122.44071982406115,
-        37.7546745592597
-      ],
-      [
-        -122.44044034714439,
-        37.75508766756418
-      ],
-      [
-        -122.44012846710689,
-        37.75567049861364
-      ],
-      [
-        -122.44005961047523,
-        37.755907472562626
-      ],
-      [
-        -122.44015276944751,
-        37.75640703677898
-      ],
-      [
-        -122.44018924596475,
-        37.756528133045606
-      ],
-      [
-        -122.44033181327828,
-        37.75666225134219
-      ],
-      [
-        -122.44070537573275,
-        37.75680778368681
-      ],
-      [
-        -122.44023039641965,
-        37.75725150991822
-      ],
-      [
-        -122.43988427866199,
-        37.757200599225015
-      ],
-      [
-        -122.4395582840298,
-        37.75708923196163
-      ],
-      [
-        -122.43904715664357,
-        37.757098777733574
-      ],
-      [
-        -122.4390310581432,
-        37.75743287897778
-      ],
-      [
-        -122.43797228289426,
-        37.757496186354615
-      ],
-      [
-        -122.43690020835358,
-        37.75758520091664
-      ],
-      [
-        -122.43582256860111,
-        37.75763591465913
-      ],
-      [
-        -122.43356465673874,
-        37.75773734204003
-      ],
-      [
-        -122.43248701698633,
-        37.7578184838447
-      ],
-      [
-        -122.4326537945671,
-        37.75940073124414
-      ],
-      [
-        -122.43275642692444,
-        37.76103365621418
-      ],
-      [
-        -122.43285905928187,
-        37.76182474975916
-      ],
-      [
-        -122.43293905683697,
-        37.762663828581225
-      ],
-      [
-        -122.43313584578942,
-        37.764194972084184
-      ],
-      [
-        -122.43195511207469,
-        37.76511200542191
-      ],
-      [
-        -122.4310229538788,
-        37.76584889897153
-      ],
-      [
-        -122.42984222016415,
-        37.766757724237436
-      ],
-      [
-        -122.4285786279431,
-        37.76783028743526
-      ],
-      [
-        -122.42690074319054,
-        37.76905839413027
-      ],
-      [
-        -122.42561643634292,
-        37.77008998799232
-      ],
-      [
-        -122.42458070501422,
-        37.77093326434769
-      ],
-      [
-        -122.42341032861272,
-        37.77189114912355
-      ],
-      [
-        -122.42198101937899,
-        37.77302094598606
-      ],
-      [
-        -122.42074849909775,
-        37.773970616950706
-      ],
-      [
-        -122.41969205314248,
-        37.774813849048584
-      ],
-      [
-        -122.41863560718714,
-        37.77560795203675
-      ],
-      [
-        -122.41794166719687,
-        37.77623831637716
-      ]
-    ]
-  }
-]
