+++ examples/additional-overlays/src/choropleth-overlay.md
