+++ examples/additional-overlays/webpack.config.js
@@ -19,7 +19,7 @@ const config = {
   mode: 'development',
 
   entry: {
-    app: resolve('./app.js')
+    app: resolve('./src/app.js')
   },
 
   output: {
